Refactor packages in api
