+++ api/src/main/java/org/hyperledger/merkletree/MerkleTree.java
@@ -11,7 +11,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hyperledger.common;
+package org.hyperledger.merkletree;
+
+import org.hyperledger.common.Hash;
 
 import java.util.ArrayList;
 import java.util.List;
