+++ api/src/main/java/org/hyperledger/merkletree/MerkleTreeNode.java
@@ -11,20 +11,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hyperledger.common;
+package org.hyperledger.merkletree;
 
-import java.io.IOException;
-import java.time.LocalTime;
+import org.hyperledger.common.Hash;
 
-public interface Header {
-    BID getID();
+/**
+ * A node in the merkle tree.
+ * It holds the id of a transaction at the leafs and digest of digests on higher level
+ */
+public interface MerkleTreeNode {
+    /**
+     * The ID of the transaction if a leaf node otherwise digest of digests
+     *
+     * @return
+     */
+    Hash getID();
 
-    BID getPreviousID();
-
-    MerkleRoot getMerkleRoot();
-
-    @Deprecated
-    int getCreateTime();
-
-    LocalTime getLocalCreateTime();
+    /**
+     * The height in the tree. 0 is leaf.
+     *
+     * @return
+     */
+    int getMerkleHeight();
 }
