+++ api/src/test/java/org/hyperledger/api/connector/GRPCClientTest.java
@@ -11,13 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.hyperledger.connector;
+package org.hyperledger.api.connector;
 
 import org.hyperledger.api.HLAPIException;
 import org.hyperledger.api.HLAPITransaction;
-import org.hyperledger.common.*;
+import org.hyperledger.transaction.TID;
+import org.hyperledger.transaction.Transaction;
 import org.junit.Before;
-import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
